requirements.txt
python-dotenv
fastapi
uvicorn[standard]

pydantic
numpy

python-dotenv
openai




